fix: set things up correctly
Move configuration files around and set things up so that things work
correctly.
