--- package-fireproof-core.json
@@ -1,77 +1,0 @@
-{
-  "name": "@fireproof/core",
-  "version": "must-set",
-  "description": "Live ledger for the web.",
-  "type": "module",
-  "module": "./index.js",
-  "main": "./index.cjs",
-  "types": "./index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./index.js",
-      "browser": "./index.js",
-      "require": "./index.cjs",
-      "types": "./index.d.ts"
-    },
-    "./node": {
-      "import": {
-        "browser": "./web/index.js",
-        "default": "./node/index.js"
-      },
-      "browser": "./web/index.js",
-      "require": "./node/index.cjs",
-      "types": "./node/index.d.ts"
-    },
-    "./deno": {
-      "import": {
-        "browser": "./web/index.js",
-        "default": "./deno/index.js"
-      },
-      "browser": "./web/index.js",
-      "require": "./deno/index.cjs",
-      "types": "./deno/index.d.ts"
-    },
-    "./web": {
-      "import": "./web/index.js",
-      "browser": "./web/index.js",
-      "require": "./web/index.cjs",
-      "types": "./web/index.d.ts"
-    },
-    "./react": {
-      "import": "./react/index.js",
-      "browser": "./react/index.js",
-      "require": "./react/index.cjs",
-      "types": "./react/index.d.ts"
-    }
-  },
-  "scripts": {},
-  "keywords": ["ledger", "JSON", "document", "IPLD", "CID", "IPFS"],
-  "contributors": ["J Chris Anderson", "Alan Shaw", "Travis Vachon", "Mikeal Rogers", "Meno Abels"],
-  "author": "J Chris Anderson",
-  "license": "Apache-2.0 OR MIT",
-  "homepage": "https://use-fireproof.com",
-  "gptdoc": "import { fireproof } from 'use-fireproof; const db = fireproof('app-db-name'); const ok = await db.put({ anyField: ['any','json'] }); const doc = await db.get(ok.id); await db.del(doc._id); useEffect(()=> db.subscribe(updateReactStateAsyncFn), []); const result = await db.query('anyField', {range : ['a', 'z']}); result.rows.map(({ key }) => key);",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/fireproof-storage/fireproof.git"
-  },
-  "bugs": {
-    "url": "https://github.com/fireproof-storage/fireproof/issues"
-  },
-  "devDependencies": {},
-  "peerDependencies": {
-    "react": "from-package-json"
-  },
-  "dependencies": {
-    "@adviser/cement": "from-package-json",
-    "@fireproof/vendor": "from-package-json",
-    "multiformats": "from-package-json",
-    "@ipld/unixfs": "from-package-json",
-    "charwise": "from-package-json",
-    "prolly-trees": "from-package-json",
-    "idb": "from-package-json",
-    "p-limit": "from-package-json",
-    "p-map": "from-package-json",
-    "p-retry": "from-package-json"
-  }
-}
