+++ packages/core/setup.file.js
