+++ packages/core/setup.memory.js
