+++ packages/core/vitest.file.config.ts
@@ -1,12 +1,9 @@
 import { defineConfig } from "vitest/config";
 
-import tsconfigPaths from "vite-tsconfig-paths";
-
 export default defineConfig({
-  plugins: [tsconfigPaths()],
   test: {
     name: "file",
-    exclude: ["tests/react/**", "examples/**", "tests/gateway/indexdb"],
+    exclude: ["tests/gateway/indexdb"],
     include: ["tests/**/*test.?(c|m)[jt]s?(x)"],
     globals: true,
     setupFiles: "./setup.file.js",
