+++ packages/use-fireproof/eslint.config.mjs
@@ -1,0 +1,9 @@
+import tseslint from "typescript-eslint";
+
+import shared from "../../eslint.shared.mjs";
+
+const opts = tseslint.config(
+  ...shared,
+);
+
+export default opts;
