+++ packages/use-fireproof/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../tsconfig.shared.json",
+}
