+++ packages/use-fireproof/vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    environment: "jsdom",
+    include: ["tests/**/*test.?(c|m)[jt]s?(x)"]
+  }
+});
